requirements.txt
numpy==1.19.5
tensorflow==1.13.1
fastapi==0.63.0
Pillow==8.1.2
pymongo==3.11.3
scikit_learn==0.24.1
